OMD-1374: extract processOcrJobAsync; dispose ocrLegacy.js (1889 LOC)
Phase 7.2 — finishes the deferred follow-up to OMD-1370 (Tier 4 cleanup).

The audit flagged routes/ocrLegacy.js as 1889 LOC of mostly-dead Express
handlers held alive by ONE actively-used function: processOcrJobAsync
(lines 971-1407, ~437 LOC). The legacy /api/ocr/* router itself was not
mounted anywhere — only the function was consumed.

Changes:
- New module: server/src/ocr/processOcrJobAsync.js — clean home for the
  function, with only the imports it actually uses (path, fs.promises,
  isDist helper). Function signature and behavior unchanged.
- server/src/routes/ocr/index.ts: removed the re-export of
  processOcrJobAsync (was: const ocrLegacy = require('../ocrLegacy');
  export const processOcrJobAsync = ocrLegacy.processOcrJobAsync). No
  other consumer needed it.
- server/src/workers/ocrFeederWorker.ts: now imports processOcrJobAsync
  directly from server/src/ocr/processOcrJobAsync, eliminating the
  indirection through routes/ocr/index.ts. Error message updated to
  reflect the new path.
- server/src/routes/ocrLegacy.js: moved to /var/www/om-disposal/2026-05-02-ocrLegacy-extraction/
  with a README explaining provenance. The 970 dead-router lines are
  finally retired.

Verification:
- grep -rn 'ocrLegacy' server/src returns only the provenance comment
  in the new file
- All processOcrJobAsync references resolve to the new module
- Function signature unchanged: (db, jobId, imagePath, options)
- om-ocr-worker systemd service must be restarted after deploy
  (handled by om-deploy.sh)
