+++ server/src/routes/ocr/index.ts
@@ -84,7 +84,3 @@ export function mountOcrRoutes(app: any, upload: any) {
 
   console.log('✅ [OCR] All OCR routes mounted (admin + 7 church-scoped modules)');
 }
-
-// Re-export processOcrJobAsync from legacy module for feeder worker compatibility
-const ocrLegacy = require('../ocrLegacy');
-export const processOcrJobAsync = ocrLegacy.processOcrJobAsync;
